fix/udp: fix timeouts
When response timesout after messages (syn, ack, etc.), we used to retry the message and extend the previous timeout by 200 miliseconds.
But in reality because of network latency and CPU execution time we might leave only a fraction of time for the future message. Or even worse case, we might specify timeout that is already finished.
This little change makes sure we give at least 200 miliseconds for each response to arrive.