fix/sockets: numerous fixes
1. UDP 'connected' sockets in BSD compliant behaviour will receive ICMP
errors. This leads to non-standard zero-byte read on Windows either due
to Windows or mio. This case is now handled and we progress ttl-runners
further.
2. On Linux/Mac ICMP port not reachable results in Connection-refused
error on write, again for a 'connected' UDP socket. This in now handled.
3. Edged triggered notifications are not respected on Windows due to
some reason, if the TCP socket has errored out. This results in error
flooding due to continuous trigger of ready handler with an error
condition as if it were Level triggered. This is now handled and early
socket deregisteration in TCP Puncher is done before connect-retry.
4. Defensive cancelation of timers to not leave about zombie timers in
some unknown edge-case - in TCP Puncher.
5. Supply nat-type information even in case of Rendezvous failure.
6. UDP Puncher logs are enhanced.
7. Test cases and examples are updated to reflect the above.
8. Trivial tweak to peer-example to copy our rendezvous info to the
clipboard if possible, on Linux.