fix/ci: skip rustfmt on Windows
We hit the rustfmt bug: https://github.com/rust-lang/rustfmt/issues/1873 .
Currently, rustfmt fails to format when path attribute has relitive
paths.
Usually, our code does not have too much windows specific code and in
this case there's no such code at the moment. So, it's the easiest fix
is to disable rustfmt on Windows for now hoping for the fix in the future.